UBUNTU: [packaging] handle downloads from the librarian better
When downloading from the librarian on a buildd (the common case) we cannot
use https to obtain those files (we are only guarenteed access to the
librarian on http from the DCs).  Add a new downloader which understands
this and fixed up downloads triggered via librariant +files redirects.
Switch us from wget to curl for downloads under the hood so update the
build-depends.

BugLink: http://bugs.launchpad.net/bugs/1850958
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>