UBUNTU: SAUCE: shiftfs: fix dentry revalidation
BugLink: https://bugs.launchpad.net/bugs/1879196

While fixing [1] we introduced a regression and triggered [2] whereby
the underlay and overlay go out of sync because they disagree about the
dcache. Such situation easily arise when the underlay is modified
directly. This means the overlay needs to revalidate the dentry in the
dcache to catchup with the underlay.

Note, in general it's not advisable to directly modify the underlay
while a shiftfs mount is on top. In some way this means we need to keep
two caches in sync and it's hard enough to keep a single cache happy.
But shiftfs' use-case is inherently prone to be used for exactly that.
So this is something we have to navigate carefully and honestly we have
no full model upstream that does the same. Overlayfs has the copy-up
behavior which let's it get around most of the issues but we don't have
it and ecryptfs is broken in such scenarios which we verified quite a
while back.
In any case, I built a kernel with this patch and re-ran all regressions
that are related to this that we have so far (cf.  [1], [2], and [3]).
None of them were reproducible with this patch here. So we still fix the
ESTALE issue but also keep underlay and overlay in sync.

[1]: https://bugs.launchpad.net/bugs/1872757
[2]: https://bugs.launchpad.net/bugs/1879196
[3]: https://bugs.launchpad.net/bugs/1860041
Cc: Seth Forshee <seth.forshee@canonical.com>
Cc: James Troup <james.troup@canonical.com>
Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>